# packages/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "用户列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDto"
                  }
                }
              }
            }
          }
        },
        "tags": ["Users"]
      },
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "tags": ["Users"]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "用户详情",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "tags": ["Users"]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "tags": ["Users"]
      },
      "delete": {
        "operationId": "UsersController_deactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "停用成功"
          }
        },
        "tags": ["Users"]
      }
    },
    "/users/{id}/reset-password": {
      "post": {
        "operationId": "UsersController_resetPassword",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "临时密码（仅此一次展示）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": ["Users"]
      }
    },
...
    "/catalog/{dimension}": {
      "get": {
        "operationId": "CatalogController_listByDimension",
        "parameters": [
          {
            "name": "dimension",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "某维度全部选项（表单过滤停用项，历史值仍可展示）"
          }
        },
        "tags": ["catalog"]
      }
    },
    "/catalog": {
      "get": {
        "operationId": "CatalogController_listAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "全部字典项（含停用，admin 维护用）"
          }
        },
        "tags": ["catalog"]
      },
      "post": {
        "operationId": "CatalogController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDimensionOptionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "新增字典项"
          }
        },
        "tags": ["catalog"]
      }
    },
    "/catalog/{id}": {
      "patch": {
        "operationId": "CatalogController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDimensionOptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新字典项"
          }
        },
        "tags": ["catalog"]
      },
      "delete": {
        "operationId": "CatalogController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "停用字典项（软删）"
          }
        },
        "tags": ["catalog"]
      }
    },
...
    "/expenses": {
      "post": {
        "operationId": "ExpensesController_upsert",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExpenseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "保存/更新当日费用（upsert）"
          }
        },
        "tags": ["expenses"]
      },
      "get": {
        "operationId": "ExpensesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "我的费用（可按月 YYYY-MM 筛选）"
          }
        },
        "tags": ["expenses"]
      }
    },
    "/expenses/{id}/submit": {
      "post": {
        "operationId": "ExpensesController_submit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseCommandDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "提交（计入统计）"
          }
        },
        "tags": ["expenses"]
      }
    },
    "/expenses/{id}/void": {
      "post": {
        "operationId": "ExpensesController_void",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseCommandDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "作废（剔除统计，留痕）"
          }
        },
        "tags": ["expenses"]
      }
    },
...
  },
  "info": {
    "title": "Lite CRM API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "Role": {
        "type": "string",
        "enum": ["sales", "executive", "assistant", "admin"],
        "description": "角色"
      },
...
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "用户 ID"
          },
          "username": {
            "type": "string",
            "description": "登录用户名"
          },
          "displayName": {
            "type": "string",
            "description": "显示名"
          },
          "role": {
            "description": "角色",
            "allOf": [
              {
                "$ref": "#/components/schemas/Role"
              }
            ]
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "description": "手机号（钉钉绑定预留）"
          },
          "reportsToId": {
            "type": "string",
            "nullable": true,
            "description": "直属上级 ID（组织树）"
          },
          "region": {
            "type": "string",
            "nullable": true,
            "description": "区域"
          },
          "isActive": {
            "type": "boolean",
            "description": "是否启用"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "创建时间"
          },
          "version": {
            "type": "number",
            "description": "当前版本号"
          }
        },
        "required": ["id", "username", "displayName", "role", "isActive", "createdAt", "version"]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "登录用户名"
          },
          "password": {
            "type": "string",
            "description": "初始密码（至少 8 位）"
          },
          "displayName": {
            "type": "string",
            "description": "显示名"
          },
          "role": {
            "description": "角色",
            "allOf": [
              {
                "$ref": "#/components/schemas/Role"
              }
            ]
          },
          "reportsToId": {
            "type": "string",
            "description": "直属上级 ID（组织树）"
          },
          "phone": {
            "type": "string",
            "description": "手机号（钉钉绑定预留）"
          },
          "region": {
            "type": "string",
            "description": "区域"
          }
        },
        "required": ["username", "password", "displayName", "role"]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "number",
            "description": "用户当前版本号，用于防止并发覆盖"
          },
          "displayName": {
            "type": "string",
            "description": "显示名"
          },
          "role": {
            "description": "角色",
            "allOf": [
              {
                "$ref": "#/components/schemas/Role"
              }
            ]
          },
          "reportsToId": {
            "type": "string",
            "nullable": true,
            "description": "直属上级 ID（组织树，null=顶层）"
          },
          "phone": {
            "type": "string",
            "description": "手机号（钉钉绑定预留）"
          },
          "region": {
            "type": "string",
            "description": "区域"
          },
          "isActive": {
            "type": "boolean",
            "description": "是否启用"
          }
        },
        "required": ["version"]
      },
...
      "CustomerDimension": {
        "type": "string",
        "enum": [
          "industry",
          "sub_industry",
          "customer_type",
          "product_line",
          "source",
          "complaint_type",
          "trade_type",
          "opportunity_source",
          "visit_type",
          "contact_function"
        ],
        "description": "维度"
      },
      "CreateDimensionOptionDto": {
        "type": "object",
        "properties": {
          "dimension": {
            "description": "维度",
            "allOf": [
              {
                "$ref": "#/components/schemas/CustomerDimension"
              }
            ]
          },
          "name": {
            "type": "string",
            "description": "稳定字典值（维度内唯一，如 new_customer）"
          },
          "label": {
            "type": "string",
            "description": "展示名称（如新客户开发）"
          },
          "sortOrder": {
            "type": "number",
            "description": "排序权重"
          }
        },
        "required": ["dimension", "name", "label"]
      },
      "UpdateDimensionOptionDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "number",
            "description": "字典项当前版本号，用于防止并发覆盖"
          },
          "label": {
            "type": "string",
            "description": "展示名称"
          },
          "sortOrder": {
            "type": "number",
            "description": "排序权重"
          },
          "isActive": {
            "type": "boolean",
            "description": "是否启用"
          }
        },
        "required": ["version"]
      },
...
      "CreateExpenseDto": {
        "type": "object",
        "properties": {
          "expenseDate": {
            "type": "string",
            "description": "费用日期"
          },
          "version": {
            "type": "number",
            "description": "更新已有草稿时的版本号；新建时不传"
          },
          "tobaccoAlcohol": {
            "type": "number",
            "description": "烟酒"
          },
          "gifts": {
            "type": "number",
            "description": "礼品"
          },
          "dining": {
            "type": "number",
            "description": "餐叙"
          },
          "entertainment": {
            "type": "number",
            "description": "招待"
          },
          "lodging": {
            "type": "number",
            "description": "住宿"
          },
          "notes": {
            "type": "string",
            "description": "备注"
          }
        },
        "required": ["expenseDate"]
      },
      "ExpenseCommandDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "number",
            "description": "当前费用版本号，用于防止并发覆盖"
          }
        },
        "required": ["version"]
      },
...
    }
  }
}